Reorganize npm package content
